requirements.txt
torch>=1.0.0
torchvision
Pillow
numpy